added spacy
