+++ examples/kubeflow/pipeline/spacy_tokenize/.s2i/environment
@@ -1,0 +1,4 @@
+MODEL_NAME=Transformer
+API_TYPE=REST
+SERVICE_TYPE=MODEL
+PERSISTENCE=0
