+++ examples/kubeflow/pipeline/spacy_tokenize/build_image.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+s2i build . seldonio/seldon-core-s2i-python3:0.6 spacy_tokenizer:0.1
+
